app/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
fastapi==0.116.1
filelock==3.19.1
flake8==7.3.0
flake8-bugbear==24.12.12
flake8_import_order==0.19.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
virtualenv==20.34.0
alembic
psycopg2
